feat: make sidebar editor awesome again
